+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-aniso8601 8.0.1-dev.0
+aniso8601 8.1.0-dev.0
 =====================
 
 *Release date: YYYY-MM-DD*